fix: added check for buffer
After suggestion by Michal.

fix the bug with the buffer

The offset and the length of the newly read part of the file were not taken into account causing old text to be preserved at the end of buffer. 

That is because shorter lines that were read after longer lines were stored in the buffer previously. 

It is now fixed by taking into account the offset (0) and the read length.

Therefore the code has also been simplified by removed "synchronized" and "AtomicReference", which were replaced by a simple lock.